src/requirements.txt
aws-cdk-lib
constructs
cfn-lint
pytest
pytest-snapshot